backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.30.0
e2b-code-interpreter==1.0.3
e2b==1.0.5
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
python-dotenv>=1.0.1
